// 01-ng-14/projects/sadad-cheque-ui-component/src/lib/demo-viewer/demo-table/models/demo-table-model.ts
export class DemoTableModel {
  position?: number;
  myImageIcon?: string;
  myCssIcon?: string;
  requestDate?: Date;
  accountNumber?: number;
  accountType?: string;
  paperCount?: number;
  office?: number;
  branchCode?: number;
  printCenter?: number;
  customerId?: number;
  requestEntrance?: string;
  status?: string;
  lastActionDate?: Date;
  actor?: string;
  shamsiDateNumber?: number;
}